refactores components adds .svg logo
