+++ src/styles/globals.scss
@@ -152,7 +152,7 @@ input[type='number']::-webkit-outer-spin-button {
     }
   }
 
-  .logo {
+  .logo {  
     height: 100px;
   }
